wip skeleton query service
