--- apiserver/.gitignore
@@ -1,2 +1,0 @@
-/deployment.yaml.out
-/jupyter_notebook_config.py
