--- apiserver/deployment.yaml
@@ -1,63 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: apiserver
-spec:
-  selector:
-    matchLabels:
-      app: apiserver
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        app: apiserver
-    spec:
-{% if deploy %}
-      priorityClassName: production
-{% endif %}
-      containers:
-        - name: apiserver
-          image: "{{ apiserver_image.image }}"
-          env:
-           - name: HAIL_AUTH_URL
-             value: "http://auth.{{ default_ns.name }}"
-          ports:
-           - containerPort: 5000
-          volumeMounts:
-           - mountPath: /hail-vdc-sa-key
-             name: hail-vdc-sa-key
-           - name: session-secret-keys
-             mountPath: /session-secret-keys
-             readOnly: true
-          resources:
-            requests:
-              memory: "3.75G"
-              cpu: "0.9"
-          readinessProbe:
-            httpGet:
-              path: /healthcheck
-              port: 5000
-            initialDelaySeconds: 5
-            periodSeconds: 5
-      volumes:
-        - name: hail-vdc-sa-key
-          secret:
-            secretName: hail-vdc-sa-key
-        - name: session-secret-keys
-          secret:
-            secretName: session-secret-keys
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: apiserver
-  labels:
-    app: apiserver
-spec:
-  ports:
-  - name: hail
-    port: 5000
-    protocol: TCP
-    targetPort: 5000
-  selector:
-    app: apiserver
