+++ hail/src/main/scala/is/hail/backend/ServiceBackend.scala
@@ -1,0 +1,11 @@
+package is.hail.backend
+
+object ServiceBackend {
+  def apply(): ServiceBackend = {
+    new ServiceBackend()
+  }
+}
+
+class ServiceBackend() {
+  def request(): Int = 5
+}
