+++ query/.gitignore
@@ -1,0 +1,2 @@
+/Dockerfile.out
+/deployment.yaml.out
