+++ query/setup.cfg
